First commit of q_reverse_k()
If 'head' is NULL, or the number of node in the queue is 0 or 1,
just return.
Use a variable 'count' to record how many nodes have been visited,
and when the remainder of dividing k is 'k-1', we need to move the list.
Since we can't use 'malloc' in this function, we can only use 'LIST_HEAD'
macro to initialize another list head. Then, we use 'list_cut_position()'
to move current k nodes to 'head_to', and reverse that list by
'q_reverse()'. Finally, we move the list back to the original list.