Update q_insert_tail by using q_insert_head
To increase readability and reusability, just use q_insert_head to
implement q_insert_tail.
First, checkout whether head is NULL or not. Once the head is not NULL,
the prev pointer of head would not be NULL. Then, we can call
q_insert_head and replace arument from head to head->prev.