Update q_remove_tail by using q_remove head
To increase readibility and reusability of code. First, checkout head is
NULL or the list is empty. This step is the same as original version,
but it is not enough. Moreover, we need to check the if there is only
one node in the list.If so, then remove_head is just like remove_tail.
Otherwise, we can replace argumnet from head to head->prev->prev and
call q_remove_head.